pyproject.toml
[project]
name = "bank-webhook-processor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "ruff>=0.11.11",
]
